requirements-job.txt
wheel
google-cloud-bigquery==3.0.0
google-cloud-secret-manager==2.22.0
google-auth==2.22.0
google-auth-oauthlib==0.4.6
google-api-python-client==2.70.0
numpy==1.26.4
pandas==1.3.3
pandas-gbq==0.17.5
pyarrow==5.0.0
python-dotenv==1.0.0
pytz==2023.3
requests==2.28.1
urllib3>=1.26.5
slack-sdk==3.27.1
psycopg2-binary==2.9.7 